object-detection: address PR review - guard intrinsics+dims together, move is_inference_stream to stream.h
